+++ test/unit/test_graph_selector_methods.py
@@ -470,7 +470,6 @@ def manifest(seed, source, ephemeral_model, view_model, table_model, ext_source,
         docs={},
         files={},
         reports={},
-        generated_at=datetime.utcnow(),
         disabled=[],
     )
     return manifest