+++ test/unit/test_parser.py
@@ -860,7 +860,7 @@ class ProcessingTest(BaseParserTest):
             self.doc.unique_id: self.doc,
         }
         self.manifest = Manifest(
-            nodes=nodes, sources=sources, macros={}, docs=docs, disabled=[], files={}, reports={}, generated_at=mock.MagicMock()
+            nodes=nodes, sources=sources, macros={}, docs=docs, disabled=[], files={}, reports={}
         )
 
     def test_process_docs(self):